2.04.0:minrog:7 Fix for issue #86
The main 55000 option was popping up along with a child options.  I hid
the main window and swapped the previous execution events into an
immediate wrapper.  Tested with many scenarios of vassal type and
ambition flag, should work the way it was supposed to now.  Testing and
feedback is good though.

Future Plan:
Count -> Duke -> King -> Emperor
All 4 should be able to finish the culture_conversion ambition, but
right now it is set to top liege and personal demesne owner.  I'll get
back to it eventually because, AI.