2.04.0:schwarherz:20 celtic stuff
corrected some of the celtic references. There shouldn't be any more
references to druidic, druidic_reformed, or gaelic_pagan_reformed. Also,
it shouldn't be possible to convert to brythonic or gaelic pagan from
any religion but celtic_pagan_reformed. What is referenced where will
need played with a little bit and I'm not sure if brythonic or gaelic
pagan are referenced in heresy events yet so they may not form at all.